requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1   
python-dotenv==0.14.0
pymongo==3.10.1
Werkzeug==1.0.1
